Added comments to the code

+++ eva_data_analysis.py
@@ -6,9 +6,11 @@ input_file = open('./eva-data.json', 'r', encoding='ascii')
 output_file = open('./eva-data.csv', 'w', encoding='utf-8')
 graph_file = './cumulative_eva_graph.png'
 
+# Reaad in the data from a JSON file into pandas data frame
 eva_df = pd.read_json(input_file, convert_dates=['date'])
-eva_df['eva'] = eva_df['eva'].astype(float)
-eva_df.dropna(axis=0, inplace=True)
+eva_df['eva'] = eva_df['eva'].astype(float) # Convert eva values to float
+# Clean the data to removing incomplete rows and sorting by date
+eva_df.dropna(axis=0, inplace=True) # Drop rows that have any missing data
 eva_df.sort_values('date', inplace=True)
 
 eva_df.to_csv(output_file, index=False)
@@ -18,6 +20,6 @@ eva_df['cumulative_time'] = eva_df['duration_hours'].cumsum()
 plt.plot(eva_df['date'], eva_df['cumulative_time'], 'ko-')
 plt.xlabel('Year')
 plt.ylabel('Total time spent in space to date (hours)')
-plt.tight_layout()
+plt.tight_layout() # Possibly adjust  margins
 plt.savefig(graph_file)
 plt.show()
